:test_tube: test: storybook 수정
